requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
requests==2.32.4
python-dotenv==1.0.1
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
PyMuPDF==1.26.3
python-docx==1.1.2
numpy==2.3.2
